Refresh logs after call
